Handle failure to ACK addr on chip wakeup
The APDS9500 requires a non-standard i2c "wakeup" command from a cold start.

This entails writing its i2c address (0x73) to the bus. However it will not ACK this byte, resulting in an IO exception.

This patch attempts to handle the IO exception, wait for wakeup to complete, and re-initialise the library.